requirements.txt
tensorflow==2.18.0
opencv-python==4.11.0.86
numpy==1.26.4
scikit-learn==1.2.2
Pillow==11.1.0
ipywidgets==8.1.5
matplotlib